# Optimize json_parse with cast