Json objects are stored as strings in database in many use cases. It is inefficient to convert a string into json and then convert the json into an array or map. The patch optimize the two steps into way.

TPCH benchmark shows a CPU save around **30% - 40%**

100 entries in an array
```
before sql_cast_json_parse :: 1012.584 cpu ms ::    0B peak memory :: in   15K,      0B,   14.8K/s,      0B/s :: out   15K,  7.22MB,   14.8K/s,  7.13MB/s
after  sql_cast_json_parse ::  593.713 cpu ms ::    0B peak memory :: in   15K,      0B,   25.3K/s,      0B/s :: out   15K,  7.22MB,   25.3K/s,  12.2MB/s
```


10 enties in an array
```
before sql_cast_json_parse ::  162.843 cpu ms ::    0B peak memory :: in   15K,      0B,   92.1K/s,      0B/s :: out   15K,   806KB,   92.1K/s,  4.83MB/s
after  sql_cast_json_parse ::  113.290 cpu ms ::    0B peak memory :: in   15K,      0B,    132K/s,      0B/s :: out   15K,   806KB,    132K/s,  6.94MB/s
```

The PR resolves https://github.com/prestodb/presto/issues/8286