@wenleix 2017-07-01
A minor comment for the commit message:

```
Json objects are stored as strings in database in many use cases. It is
inefficient to convert a string into json and then convert the json into
an array or map. The patch optimize the two steps into way.
```

I think it's ok to convert a string into `JSON`. In my opinion, the root cause is today we don't have an efficient internal representation for `JSON` type. A reformatted string is used to represented `JSON` internally, which makes the current execution to parse twice.