@highker 2017-07-24
Comments addressed; blocked by https://github.com/prestodb/presto/pull/8217